+++ src/poolcore/api.h
@@ -3,4 +3,4 @@
 
 class p2pPeer;
 
-void poolcoreRequestHandler(p2pPeer *peer, uint64_t id, void *buffer, size_t size, void *arg);
+void poolcoreRequestHandler(p2pPeer *peer, uint32_t id, void *buffer, size_t size, void *arg);
